% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_data-class.R
\name{add_integrated_data}
\alias{add_integrated_data}
\alias{is.integrated_data}
\alias{print.integrated_data}
\alias{plot.integrated_data}
\alias{summary.integrated_data}
\title{An integrated data object}
\usage{
add_integrated_data(data, process_model, process_link,
  observation_model = "naive")

is.integrated_data(model)

\method{print}{integrated_data}(x, ...)

\method{plot}{integrated_data}(x, ...)

\method{summary}{integrated_data}(object, ...)
}
\arguments{
\item{data}{something}

\item{observation_model}{something else}

\item{...}{additional arguments}
}
\value{
An object of class \code{integrated_data}, which contains information on the data module and
  can be passed to a \link[integrated]{integrated_model}
}
\description{
A \code{integrated_model} call adds a component data object to process model
  created with \link[integrated]{define_integrated_process}
}
\examples{

library(integrated)

# prepare an example model
data <- define_integrated_data()
                        
\dontrun{                 
# summarise data module
model
summary(model)
plot(model)
}

# check if an object is an integrated_data object
  
\dontrun{
is.integrated_data(model)
}

# Print information about an 'integrated_data' object

\dontrun{
print(x)
}

# Plot an 'integrated_data' object

\dontrun{
plot(x)
}

# Summarise an 'integrated_data' object

\dontrun{
summary(x)
}
}
