% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_process-class.R
\name{integrated_process}
\alias{integrated_process}
\alias{is.integrated_process}
\alias{print.integrated_process}
\alias{summary.integrated_process}
\title{An integrated process object}
\usage{
integrated_process(type, classes, structure = "stage",
  density_dependence = "none", replicates = 1, params = list())

is.integrated_process(model)

\method{print}{integrated_process}(x, ...)

\method{summary}{integrated_process}(object, ...)
}
\arguments{
\item{type}{something}

\item{classes}{something}

\item{structure}{something}

\item{density_dependence}{something}

\item{replicates}{number of distinct populations}

\item{...}{additional arguments}
}
\value{
An object of class \code{integrated_process}, which can be passed to
   \link[integrated]{integrated_data} and \link[integrated]{integrated_model}
}
\description{
An \code{integrated_process} object contains the underlying process model
 for an integrated population analysis
}
\examples{

library(integrated)

# prepare an example model
mpm_process <- integrated_process(type = 'MPM')

# check if an object is an integrated_process object
  
\dontrun{
is.integrated_process(model)
}

# Print information about an 'integrated_process' object

\dontrun{
print(x)
}

# Summarise an 'integrated_process' object

\dontrun{
summary(x)
}
}
