% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_model-class.R
\name{build_integrated_model}
\alias{build_integrated_model}
\alias{is.integrated_model}
\alias{print.integrated_model}
\alias{plot.integrated_model}
\alias{summary.integrated_model}
\title{An integrated model object}
\usage{
build_integrated_model(integrated_process, ...)

is.integrated_model(model)

\method{print}{integrated_model}(x, ...)

\method{plot}{integrated_model}(x, ...)

\method{summary}{integrated_model}(object, ...)
}
\arguments{
\item{...}{\code{integrated_data} objects created with \link[integrated]{define_integrated_data}}
}
\value{
An object of class \code{integrated_model}, which can be passed to
   \link[greta]{mcmc} and has associated `print`, `plot`, and `summary` methods
}
\description{
A \code{integrated_model} object contains a compiled greta model for integrated data analysis
}
\examples{

library(integrated)

# prepare an example model
model <- build_integrated_model()
                        
\dontrun{                 
# summarise fitted model
model
summary(model)
plot(model)
}

# check if an object is an integrated_model object
  
\dontrun{
is.integrated_model(model)
}

# Print information about an 'integrated_model' object

\dontrun{
print(x)
}

# Plot an 'integrated_model' object

\dontrun{
plot(x)
}

# Summarise an 'integrated_model' object

\dontrun{
summary(x)
}
}
