% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrated_process-class.R
\name{define_integrated_process}
\alias{define_integrated_process}
\alias{is.integrated_process}
\alias{print.integrated_process}
\alias{summary.integrated_process}
\title{An integrated process object}
\usage{
define_integrated_process(type, structure, classes, density_dependence,
  replicates = 1)

is.integrated_process(model)

\method{print}{integrated_process}(x, ...)

\method{summary}{integrated_process}(object, ...)
}
\arguments{
\item{...}{additional arguments}
}
\value{
An object of class \code{integrated_process}, which can be passed to
   \link[integrated]{define_integrated_data} and \link[integrated]{build_integrated_model}
}
\description{
A \code{integrated_process} object contains a component data object for integrated data analysis
}
\examples{

library(integrated)

# prepare an example model
data <- define_integrated_process(type = 'MPM')

# check if an object is an integrated_process object
  
\dontrun{
is.integrated_process(model)
}

# Print information about an 'integrated_process' object

\dontrun{
print(x)
}

# Summarise an 'integrated_process' object

\dontrun{
summary(x)
}
}
